+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Release Notes for Sprig
 
-## 1.0.4 - Unreleased
+## 1.1.0 - Unreleased
 ### Added
-- Added a `sprig.push` template variable that pushes a URL into the history stack.
+- Added a `sprig.triggerEvents` template variable that triggers client-side events.
+- Added a `sprig.pushUrl` template variable that pushes a URL into the history stack.
 
 ### Changed
 - Made minor tweaks to default playground.